requirements.txt
asgiref==3.2.10
boto3==1.14.32
botocore==1.17.32
dj-database-url==0.5.0
Django==3.0.8
django-bootstrap-modal-forms==2.0.0
django-filter==2.3.0
django-simple-history==2.11.0
django-storages==1.9.1
django-widget-tweaks==1.4.8
docutils==0.15.2
gunicorn==20.0.4
html5lib==1.1
jmespath==0.10.0
numpy==1.19.1
Pillow==7.2.0
psycopg2==2.8.5
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.46
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
